--- src/main/java/com/github/nhirakawa/swarm/model/SwarmMessageType.java
@@ -1,10 +1,0 @@
-package com.github.nhirakawa.swarm.model;
-
-public enum SwarmMessageType {
-
-  UUID,
-  PING,
-  PING_ACK,
-  PING_REQUEST
-
-}
