--- src/main/java/com/github/nhirakawa/swarm/protocol/MemberStatus.java
@@ -1,7 +1,0 @@
-package com.github.nhirakawa.swarm.protocol;
-
-public enum MemberStatus {
-  ALIVE,
-  SUSPECTED,
-  FAILED
-}
