--- src/main/resources/reference.conf
@@ -1,34 +1,0 @@
-swarm {
-
-  protocol = {
-    period = "2s"
-
-    messageTimeout = "1s"
-
-    tick = "100ms"
-
-    failureSubGroup = 2
-  }
-
-  runEntireClusterLocally = false
-
-  localNode = {
-    host = "localhost",
-    port = 8080
-  }
-
-  clusterNodes = [
-    {
-      host = "localhost",
-      port = 8080
-    },
-    {
-      host = "localhost",
-      port = 8081
-    },
-    {
-      host = "localhost",
-      port = 8082
-    }
-  ]
-}
