+++ swarm-netty/src/main/java/com/github/nhirakawa/swarm/ObjectMapperWrapper.java
