+++ swarm-netty/src/main/java/com/github/nhirakawa/swarm/concurrent/LoggingUncaughtExceptionHandler.java
