+++ swarm-netty/src/main/java/com/github/nhirakawa/swarm/dagger/IncomingQueue.java
