+++ swarm-netty/src/main/java/com/github/nhirakawa/swarm/transport/NettyFutureAdapter.java
