+++ swarm-netty/src/main/resources/banner.txt
