+++ swarm-protocol/src/main/java/com/github/nhirakawa/swarm/protocol/model/AbstractPingMessage.java
@@ -1,0 +1,20 @@
+package com.github.nhirakawa.swarm.protocol.model;
+
+import org.immutables.value.Value;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import com.github.nhirakawa.immutable.style.ImmutableStyle;
+
+@Value.Immutable
+@ImmutableStyle
+@JsonSerialize
+public abstract class AbstractPingMessage implements BaseSwarmMessage{
+
+  @Override
+  @Value.Auxiliary
+  @JsonIgnore
+  public SwarmMessageType getType() {
+    return SwarmMessageType.PING;
+  }
+}
