+++ swarm-protocol/src/main/java/com/github/nhirakawa/swarm/protocol/model/SwarmMessageType.java
@@ -1,0 +1,10 @@
+package com.github.nhirakawa.swarm.protocol.model;
+
+public enum SwarmMessageType {
+
+  UUID,
+  PING,
+  PING_ACK,
+  PING_REQUEST
+
+}
