+++ swarm-protocol/src/main/java/com/github/nhirakawa/swarm/protocol/protocol/MemberStatus.java
@@ -1,0 +1,7 @@
+package com.github.nhirakawa.swarm.protocol.protocol;
+
+public enum MemberStatus {
+  ALIVE,
+  SUSPECTED,
+  FAILED
+}
